+++ taskflowbackend/tasks/sodar_tasks.py
@@ -2,6 +2,8 @@
 
 import logging
 
+from copy import deepcopy
+
 from taskflowbackend.tasks.base_task import BaseTask
 
 
@@ -19,3 +21,27 @@ class SODARBaseTask(BaseTask):
         )
         self.name = '<SODAR> {} ({})'.format(name, self.__class__.__name__)
         self.project = project
+
+
+class TimelineEventExtraDataUpdateTask(SODARBaseTask):
+    """
+    Task for TimelineEvent extra data updating. Updates existing extra data
+    with the provided dictionary. May overwrite data in case of identical keys.
+    """
+
+    og_data = {}
+
+    def execute(self, tl_event, extra_data, *args, **kwargs):
+        # Store original data for revert
+        self.og_data = deepcopy(tl_event.extra_data)
+        data = tl_event.extra_data.copy()
+        data.update(extra_data)
+        tl_event.extra_data = data
+        tl_event.save()
+        self.data_modified = True
+        super().execute(*args, **kwargs)
+
+    def revert(self, tl_event, extra_data, *args, **kwargs):
+        if self.data_modified:
+            tl_event.extra_data = self.og_data
+            tl_event.save()